Add Jaguar CD hashing
Reuse RetroAchievements implementation, move into `DiscHasher`
